First steps to adding tuning features
I added a button to the main window that should eventually open a menu to retune individual notes.  The changes to the scripts are just a start to get the menu to show up, and not nearly functional.  I don't really know what I'm doing when I edit these scripts, just paste things from other scripts that I think will be needed.  The new scripts bring up errors that prevent the program from being built.
I am not good with C++ or programming in general, so I won't be able to complete this on my own.